# openapi.yaml
openapi: 3.0.3
info:
  title: Agent Skills Actions API
  version: "1.0.0"
servers:
  - url: https://YOUR_DOMAIN
paths:
  /health:
    get:
      operationId: health
      summary: Health check
      responses:
        "200":
          description: API health status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
  /skills:
    get:
      operationId: list_skills
      summary: List all generated skills
      responses:
        "200":
          description: Skills list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListSkillsResponse"
  /get_skill:
    post:
      operationId: get_skill
      summary: Fetch SKILL.md content for a skill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GetSkillRequest"
      responses:
        "200":
          description: SKILL.md content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetSkillResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Skill not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /get_skill_file:
    post:
      operationId: get_skill_file
      summary: Fetch a skill reference or asset file
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GetSkillFileRequest"
      responses:
        "200":
          description: File content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetSkillFileResponse"
        "400":
          description: Invalid skill_id or path
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Skill or file not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /search_docs:
    post:
      operationId: search_docs
      summary: Search references for query terms
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchDocsRequest"
      responses:
        "200":
          description: Search hits
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchDocsResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Skill not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthResponse:
      type: object
      required:
        - ok
      properties:
        ok:
          type: boolean
          example: true

    SkillInfo:
      type: object
      required:
        - skill_id
        - name
        - description
        - required_tools
      properties:
        skill_id:
          type: string
          example: branding
        name:
          type: string
          example: Branding
        description:
          type: string
          example: Umbrella brand skill for writing, reviewing, and terminology checks using available docs.
        required_tools:
          type: array
          items:
            type: string
          example:
            - get_skill_file
            - search_docs

    ListSkillsResponse:
      type: object
      required:
        - skills
      properties:
        skills:
          type: array
          items:
            $ref: "#/components/schemas/SkillInfo"

    GetSkillRequest:
      type: object
      required:
        - skill_id
      properties:
        skill_id:
          type: string
          pattern: "^[a-z0-9-]+$"
          example: branding

    GetSkillResponse:
      type: object
      required:
        - skill_id
        - content
      properties:
        skill_id:
          type: string
        content:
          type: string

    GetSkillFileRequest:
      type: object
      required:
        - skill_id
        - path
      properties:
        skill_id:
          type: string
          pattern: "^[a-z0-9-]+$"
          example: branding
        path:
          type: string
          example: references/normalized/VOICE.md

    GetSkillFileResponse:
      type: object
      required:
        - skill_id
        - path
        - content
      properties:
        skill_id:
          type: string
        path:
          type: string
        content:
          type: string

    SearchDocsRequest:
      type: object
      required:
        - skill_id
        - query
      properties:
        skill_id:
          type: string
          pattern: "^[a-z0-9-]+$"
          example: branding
        query:
          type: string
          example: tone and claims
        top_k:
          type: integer
          minimum: 1
          maximum: 20
          default: 5

    SearchHit:
      type: object
      required:
        - path
        - score
        - snippet
        - start_line
        - end_line
      properties:
        path:
          type: string
        score:
          type: integer
        snippet:
          type: string
        start_line:
          type: integer
        end_line:
          type: integer

    SearchDocsResponse:
      type: object
      required:
        - hits
      properties:
        hits:
          type: array
          items:
            $ref: "#/components/schemas/SearchHit"

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
        message:
          type: string